requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5 